@@ -5,16 +5,20 @@
     <parent>
         <artifactId>dps</artifactId>
         <groupId>ca.bc.gov.open</groupId>
-        <version>0.4.1-SNAPSHOT</version>
+        <version>0.5.0</version>
         <relativePath>../../pom.xml</relativePath>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>ca.bc.gov.open.pssg.rsbc</groupId>
     <artifactId>dfcms-ords-client</artifactId>
-    <version>0.4.1-SNAPSHOT</version>
+    <version>0.5.0</version>
 
     <properties>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
+        <maven.compiler.plugin.version>3.10.1</maven.compiler.plugin.version>
+        <maven.test.plugin.version>3.1.2</maven.test.plugin.version>
         <!-- include properties that are specific to dpsvalidationservice -->
         <default-package>ca.bc.gov.open.ords.dfcms.client</default-package>
     </properties>
