@@ -5,16 +5,21 @@
     <parent>
         <artifactId>dps</artifactId>
         <groupId>ca.bc.gov.open</groupId>
-        <version>0.4.1-SNAPSHOT</version>
+        <version>0.5.0</version>
         <relativePath>../../pom.xml</relativePath>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>ca.bc.gov.open.pssg.rsbc</groupId>
     <artifactId>dps-email-client</artifactId>
-    <version>0.4.1-SNAPSHOT</version>
+    <version>0.5.0</version>
 
     <properties>
+        <log4j2.version>2.17.1</log4j2.version>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
+        <maven.compiler.plugin.version>3.10.1</maven.compiler.plugin.version>
+        <maven.test.plugin.version>3.1.2</maven.test.plugin.version>
         <!-- include properties that are specific to dpsvalidationservice -->
         <default-package>ca.bc.gov.open.dps.email.client</default-package>
         <maven-plugin-version>1.0.0</maven-plugin-version>
